Port CyanUtil packet implementations to alpha 20

+++ CyanAPI-1.3-1.16/src/main/java/org/asf/cyan/api/internal/modkit/components/_1_16/common/network/buffer/FriendlyByteBufOutputFlow.java
@@ -9,7 +9,7 @@ public class FriendlyByteBufOutputFlow implements OutputFlow {
 	private FriendlyByteBuf buffer = new FriendlyByteBuf(Unpooled.buffer());
 	
 	@Override
-	public void write(int data) {
+	public void write(byte data) {
 		buffer.writeByte(data);
 	}
